Add gameData & helper fns for ship placing to gameLoopAuxFns
In gameLoopAuxFns.js I added a factory function for generating a
gameData object which contains all needed data throughout the course
of the game.
Also wrote two helper functions, setShipType and
setShipDirection which set which type of ship, and in which direction
it will be placed on the board. The ship type and direction can be
updated by clicking on the corresponding buttons.
Had to modify the test suite in order to test out the game data object.