Add clearBoard & placeSelectedShip method to gameBoard
Added clearBoard method for clearing the gameBoard. I plan this to
be used for erasing ship placement if the player wants to redo the
ship placing before starting.
I also added the placeSelectedShip method, which accepts the
ship type and placement direction assigned via the ship and direction
buttons. Depending on the direction, the method calls either of the
four other methods that correspond to the placement direction.