Modify clearBoard and ship placement methods in gameBoard
Modified the clearBoard method so it also changes the ships.ship.placed
property to false.
Also added another condition to the ship placement methods:
if any of the ships have a status of placed = true, skip all other
logic. Added this because problems arose when an attempt was made for
a placed ship to be placed again, but out of bounds. The squares that
were out of bounds were not defined so an error was thrown.